python_weather_producer/requirements.txt
kafka-python
avro